fix: satisfy lint and test location checks
